Perbaikan

+++ resources/views/livewire/tabel-rombel.blade.php
@@ -52,17 +52,22 @@
         </div>
     </div>
     <div class="mx-3 my-2">
-        <a href="" data-bs-toggle="modal" data-bs-target="#inputModal"
-            class="btn btn-primary active mb-2 {{ ($filterTahunAkademik === '' or $filterKelas === '' or $allow === false) ? 'disabled' : '' }}"><i
-                class='bx bx-add-to-queue'></i> Tambah</a>
-        <a href=""
-            class="btn btn-success active mb-2 {{ ($filterTahunAkademik === '' or $filterKelas === '' or $allow === false) ? 'disabled' : '' }}"
-            data-bs-toggle="modal" data-bs-target="#importModal"
-            style="background-color: rgb(0, 185, 0);border-color: rgb(0, 185, 0)"><i class='bx bxs-file-import'></i>
-            Import</a>
-        <a class="btn btn-info mb-2 text-white {{ ($filterTahunAkademik === '' or $filterKelas === '') ? 'disabled' : '' }}"
-            wire:click="export()" style="background-color: rgb(0, 143, 0);border-color: rgb(0, 143, 0)"><i
-                class='bx bxs-file-export'></i>
+        @if ($allow !== false)
+            <a href="" data-bs-toggle="modal" data-bs-target="#inputModal"
+                class="btn btn-primary active mb-2 {{ $allow === false ? 'disabled' : '' }}"><i
+                    class='bx bx-add-to-queue'></i> Tambah</a>
+            <a href="" class="btn btn-success active mb-2 {{ $allow === false ? 'disabled' : '' }}"
+                data-bs-toggle="modal" data-bs-target="#importModal"
+                style="background-color: rgb(0, 185, 0);border-color: rgb(0, 185, 0)"><i class='bx bxs-file-import'></i>
+                Import</a>
+        @else
+            <a href="" class="btn btn-primary active mb-2 disabled"><i class='bx bx-add-to-queue'></i> Tambah</a>
+            <a href="" class="btn btn-success active mb-2 disabled"
+                style="background-color: rgb(0, 185, 0);border-color: rgb(0, 185, 0)"><i class='bx bxs-file-import'></i>
+                Import</a>
+        @endif
+        <a class="btn btn-info mb-2 text-white" wire:click="export()"
+            style="background-color: rgb(0, 143, 0);border-color: rgb(0, 143, 0)"><i class='bx bxs-file-export'></i>
             Export</a>
     </div>
     <div class="col-lg-4 col-md-4 mb-2 mx-3">
@@ -92,10 +97,15 @@
                             <td>{{ $s->nisn }}</td>
                             <td>{{ $s->nama }}</td>
                             <td>
-                                <button class="btn btn-danger {{ $allow === false ? 'disabled' : '' }}"
-                                    wire:click="deleteConfirmation({{ $s->id }})"><i
-                                        class="bx bx-trash me-1"></i>
-                                    Delete</button>
+                                @if ($allow === true)
+                                    <button class="btn btn-danger"
+                                        wire:click="deleteConfirmation({{ $s->id }})"><i
+                                            class="bx bx-trash me-1"></i>
+                                        Delete</button>
+                                @else
+                                    <button class="btn btn-danger disabled"><i class="bx bx-trash me-1"></i>
+                                        Delete</button>
+                                @endif
                             </td>
                         </tr>
                     @endforeach
